cleanup: feature macros guard their defines; scanf formats match their types
Object-like feature defines (ENABLE_EXTENSION_*, _GNU_SOURCE, __USE_GNU)
gain #ifndef guards across 26 sources: a build that also selects those
extensions on the command line kept redefining them — same effective
macro, thousands of warning lines.  Guard style mirrors each file's own
indentation.

CoMD's double-precision build fed fscanf/sscanf through "%f" formats
(simulation.h's FMT1/EMT1): scanf needs the l length modifier to store
into a double, and %lf is equally valid printf usage, so one format
serves both directions; the float branch keeps %f.  The two timers.c
copies printed a u64 through %llu — the argument is cast to unsigned
long long, matching the format both print branches share.

+++ apps/CoMD/refactored/ocr/sdsc2/timers.c
@@ -32,7 +32,7 @@ void print_timers(mdtimer_t* timers_p, u32 atoms, u32 steps)
 #else
       ocrPrintf("%16s %11llu     %8.4f      %8.4f    %8.2f\n",
 #endif
-             timer_name[ii], timers_p[ii].count, total/(double)timers_p[ii].count,
+             timer_name[ii], (unsigned long long)timers_p[ii].count, total/(double)timers_p[ii].count,
              total, total/loop_time*100.0);
   }
   double rate = (double) timers_p[total_timer].total / (atoms * steps);
